+++ src/commands/toolkit/cleaner.ts
@@ -3,10 +3,24 @@ import { Commander, CommandExists } from "../../functions/cli.ts";
 import { GetSettings } from "../../functions/config.ts";
 import { CheckForPath, JoinPaths, ParsePath } from "../../functions/filesystem.ts";
 import { ColorString, LogStuff } from "../../functions/io.ts";
+import { NameProject } from "../../functions/projects.ts";
 import type { CleanerIntensity } from "../../types/config_params.ts";
 import type { SUPPORTED_NODE_LOCKFILES } from "../../types/package_managers.ts";
 import { FknError } from "../../utils/error.ts";
+import type { tRESULT } from "../clean.ts";
 
+/**
+ * Cleans a NodeJS project.
+ *
+ * @export
+ * @async
+ * @param {SUPPORTED_NODE_LOCKFILES} lockfile
+ * @param {string} projectInQuestion
+ * @param {boolean} shouldUpdate
+ * @param {boolean} shouldClean
+ * @param {("normal" | "hard" | "maxim")} intensity
+ * @returns {Promise<void>}
+ */
 export async function PerformNodeCleaning(
     lockfile: SUPPORTED_NODE_LOCKFILES,
     projectInQuestion: string,
@@ -88,6 +102,13 @@ export async function PerformNodeCleaning(
 
 // cache cleaning is global, so doing these for every project like we used to do
 // is a waste of resources
+/**
+ * Performs a hard cleanup, AKA cleans global caches.
+ *
+ * @export
+ * @async
+ * @returns {Promise<void>}
+ */
 export async function PerformHardCleanup(): Promise<void> {
     await LogStuff(
         ColorString("Time for hard-pruning! Wait patiently, please (caches can take a while to clean).", "italic"),
@@ -190,6 +211,14 @@ export async function PerformHardCleanup(): Promise<void> {
     return;
 }
 
+/**
+ * Validates the provided intensity and handles stuff like `--`.
+ *
+ * @export
+ * @async
+ * @param {string} intensity
+ * @returns {Promise<CleanerIntensity>}
+ */
 export async function ValidateIntensity(intensity: string): Promise<CleanerIntensity> {
     if (!["hard", "hard-only", "normal", "maxim", "--"].includes(intensity)) {
         throw new FknError("Cleaner__InvalidCleanerIntensity", `Provided intensity '${intensity}' is not valid.`);
@@ -219,5 +248,67 @@ export async function ValidateIntensity(intensity: string): Promise<CleanerInten
         return confirmMaxim === true ? "maxim" : defaultIntensity;
     } else {
         return defaultIntensity;
+    }
+}
+
+/**
+ * Shows a basic report.
+ *
+ * @export
+ * @async
+ * @param {{ path: string; status: string }[]} results
+ * @returns {Promise<void>}
+ */
+export async function ShowReport(results: tRESULT[]): Promise<void> {
+    // shows a report
+    await LogStuff("Report:", "chart", false);
+    const report: string[] = [];
+    for (const result of results) {
+        const name = await NameProject(result.path);
+        const status = ColorString(result.status, "bold");
+        const elapsedTime = ColorString(result.elapsedTime, "italic");
+
+        const theResult = `${name} -> ${status}, taking us ${elapsedTime}`;
+        report.push(theResult);
+    }
+    const sortedReport = report.sort((a, b) => {
+        return a
+            .toLowerCase()
+            .localeCompare(b.toLowerCase());
+    }).join("\n");
+    await LogStuff(sortedReport, undefined, false);
+    await LogStuff(
+        ColorString(`Cleaning completed at ${new Date().toLocaleString()}`, "bright-green"),
+        "tick",
+        false,
+    );
+}
+
+/**
+ * Tells you about the protection of a project.
+ *
+ * @export
+ * @param {(string | null)} protection
+ * @param {boolean} update
+ * @returns {{
+ *     doClean: boolean;
+ *     doUpdate: boolean;
+ *     preliminaryStatus?: "Fully protected" | "Cleanup protected" | "Update protected";
+ * }}
+ */
+export function ResolveProtection(protection: string | null, update: boolean): {
+    doClean: boolean;
+    doUpdate: boolean;
+    preliminaryStatus?: "Fully protected" | "Cleanup protected" | "Update protected";
+} {
+    switch (protection) {
+        case "*":
+            return { doClean: false, doUpdate: false, preliminaryStatus: "Fully protected" };
+        case "cleanup":
+            return { doClean: false, doUpdate: true, preliminaryStatus: "Cleanup protected" };
+        case "updater":
+            return { doClean: true, doUpdate: false, preliminaryStatus: "Update protected" };
+        default:
+            return { doClean: true, doUpdate: update };
     }
 }